+++ src/filecoin/FilecoinEngine.ts
@@ -126,6 +126,7 @@ export class FilecoinEngine extends CurrencyEngine<
   }
 
   async startEngine(): Promise<void> {
+    this.engineOn = true
     this.initData()
     this.initSubscriptions()
     await super.startEngine()
@@ -133,6 +134,16 @@ export class FilecoinEngine extends CurrencyEngine<
 
   async killEngine(): Promise<void> {
     await super.killEngine()
+
+    // Wait for any transaction scanning to finish before killing engine
+    await new Promise<void>(resolve => {
+      const id = setInterval(() => {
+        if (!this.isScanning) {
+          clearInterval(id)
+          resolve()
+        }
+      }, 300)
+    })
   }
 
   async clearBlockchainCache(): Promise<void> {
@@ -140,7 +151,7 @@ export class FilecoinEngine extends CurrencyEngine<
   }
 
   async resyncBlockchain(): Promise<void> {
-    await super.killEngine()
+    await this.killEngine()
     await this.clearBlockchainCache()
     await this.startEngine()
   }